# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VizTR Super Admin API",
    "version": "1.0.0",
    "description": "Super Admin Dashboard API for VizTR Architecture Visualization Platform",
    "contact": {
      "name": "VizTR Team",
      "email": "api@viztr.studio"
    }
  },
  "servers": [
    {
      "url": "https://api.viztr.studio",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/admin/bookings/{id}/approve": {
      "post": {
        "summary": "Approve booking",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "admin_notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking approved"
          }
        }
      }
    },
...
  },
...
}